use FlyoutContent
